requirements.txt
Django==4.2.13
gunicorn==22.0.0
whitenoise==6.7.0
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.5.4
bcrypt==4.1.3
markdown==3.6